+++ android/app/src/main/java/co/edgesecure/app/MainActivity.kt
@@ -1,0 +1,54 @@
+package co.edgesecure.app
+
+import android.content.pm.ActivityInfo
+import android.os.Build
+import android.os.Bundle
+import android.view.WindowManager
+import com.facebook.react.ReactActivity
+import com.facebook.react.ReactActivityDelegate
+import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
+import com.facebook.react.defaults.DefaultReactActivityDelegate
+import com.zoontek.rnbootsplash.RNBootSplash
+import expo.modules.ReactActivityDelegateWrapper
+
+class MainActivity : ReactActivity() {
+    /**
+     * Returns the name of the main component registered from JavaScript. This is used to schedule
+     * rendering of the component.
+     */
+    override fun getMainComponentName(): String = "edge"
+
+    /**
+     * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
+     * which allows you to enable New Architecture with a single boolean flags [fabricEnabled]
+     */
+    override fun createReactActivityDelegate(): ReactActivityDelegate =
+        ReactActivityDelegateWrapper(
+            this,
+            BuildConfig.IS_NEW_ARCHITECTURE_ENABLED,
+            DefaultReactActivityDelegate(this, mainComponentName, fabricEnabled)
+        )
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        // Keep the splash screen around until we are ready to hide it:
+        RNBootSplash.init(this)
+
+        // Do not pass the saved state, as required by react-native-screens:
+        super.onCreate(null)
+
+        // Hide app contents in the background:
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            setRecentsScreenshotEnabled(false);
+        }
+
+        // Lock the app to portrait mode:
+        if (resources.getBoolean(R.bool.portrait_only)) {
+            requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+        }
+    }
+
+    // Edge addition
+    override fun invokeDefaultOnBackPressed() {
+        moveTaskToBack(true)
+    }
+}