+++ src/components/data/row/CurrencyRow.tsx
@@ -17,6 +17,8 @@ import { IconDataRow } from './IconDataRow'
 interface Props {
   marginRem?: number[] | number
   nativeAmount?: string
+  /** Override user show/hide balance settings. If unset, defaults to user show/hide balance settings. */
+  hideBalance?: boolean
   showRate?: boolean
   token?: EdgeToken
   tokenId: EdgeTokenId
@@ -27,7 +29,7 @@ interface Props {
  * A view representing the data from a wallet, used for rows, cards, etc.
  */
 const CurrencyRowComponent = (props: Props) => {
-  const { marginRem, nativeAmount, showRate = false, token, tokenId, wallet } = props
+  const { marginRem, nativeAmount, hideBalance, showRate = false, token, tokenId, wallet } = props
   const { pluginId } = wallet.currencyInfo
   const { showTokenNames = false } = SPECIAL_CURRENCY_INFO[pluginId] ?? {}
   const theme = useTheme()
@@ -52,12 +54,12 @@ const CurrencyRowComponent = (props: Props) => {
   }
 
   // Balance stuff:
-  const hideBalance = useSelector(state => !state.ui.settings.isAccountBalanceVisible)
+  const hideBalanceSetting = useSelector(state => (hideBalance == null ? !state.ui.settings.isAccountBalanceVisible : hideBalance))
   const balance = useWalletBalance(wallet, tokenId)
   const icon = <CryptoIconUi4 sizeRem={2} tokenId={tokenId} walletId={wallet.id} />
   const tickerText = showRate && wallet != null ? <TickerText wallet={wallet} tokenId={tokenId} /> : null
-  const cryptoText = <CryptoText wallet={wallet} tokenId={tokenId} nativeAmount={nativeAmount ?? balance} withSymbol hideBalance={hideBalance} />
-  const fiatText = <FiatText nativeCryptoAmount={nativeAmount ?? balance} tokenId={tokenId} wallet={wallet} hideBalance={hideBalance} />
+  const cryptoText = <CryptoText wallet={wallet} tokenId={tokenId} nativeAmount={nativeAmount ?? balance} withSymbol hideBalance={hideBalanceSetting} />
+  const fiatText = <FiatText nativeCryptoAmount={nativeAmount ?? balance} tokenId={tokenId} wallet={wallet} hideBalance={hideBalanceSetting} />
 
   let displayCurrencyCode = currencyCode
   if (showTokenNames && tokenFromId != null) {