+++ src/plugins/gui/providers/paybisProvider.ts
@@ -7,10 +7,10 @@ import { SendScene2Params } from '../../../components/scenes/SendScene2'
 import { locale } from '../../../locales/intl'
 import { lstrings } from '../../../locales/strings'
 import { EdgeAsset, StringMap } from '../../../types/types'
+import { sha512HashAndSign } from '../../../util/crypto'
 import { CryptoAmount } from '../../../util/CryptoAmount'
-import { makeUuid } from '../../../util/utils'
 import { SendErrorNoTransaction } from '../fiatPlugin'
-import { FiatDirection, FiatPaymentType, SaveTxActionParams } from '../fiatPluginTypes'
+import { FiatDirection, FiatPaymentType, FiatPluginUi, SaveTxActionParams } from '../fiatPluginTypes'
 import {
   FiatProvider,
   FiatProviderApproveQuoteParams,
@@ -53,7 +53,8 @@ const allowedPaymentTypes: AllowedPaymentTypes = {
 
 const asApiKeys = asObject({
   apiKey: asString,
-  partnerUrl: asString
+  partnerUrl: asString,
+  privateKeyB64: asString
 })
 
 const asPaymentMethodId = asValue(
@@ -193,6 +194,11 @@ const asPaymentDetails = asObject({
   amount: asString
 })
 
+const asPublicRequestResponse = asObject({
+  requestId: asString,
+  oneTimeToken: asOptional(asString)
+})
+
 type PaymentMethodId = ReturnType<typeof asPaymentMethodId>
 type PaybisBuyPairs = ReturnType<typeof asPaybisBuyPairs>
 type PaybisSellPairs = ReturnType<typeof asPaybisSellPairs>
@@ -302,13 +308,13 @@ export const paybisProvider: FiatProviderFactory = {
   makeProvider: async (params: FiatProviderFactoryParams): Promise<FiatProvider> => {
     const {
       apiKeys,
-      io: { store }
+      io: { makeUuid, store }
     } = params
-    const { apiKey, partnerUrl: url } = asApiKeys(apiKeys)
+    const { apiKey, partnerUrl: url, privateKeyB64 } = asApiKeys(apiKeys)
 
     let partnerUserId = await store.getItem('partnerUserId').catch(e => undefined)
     if (partnerUserId == null || partnerUserId === '') {
-      partnerUserId = makeUuid()
+      partnerUserId = await makeUuid()
       await store.setItem('partnerUserId', partnerUserId)
     }
 
@@ -485,7 +491,7 @@ export const paybisProvider: FiatProviderFactory = {
                 },
                 partnerUserId,
                 locale: locale.localeIdentifier.slice(0, 2),
-                passwordless: false,
+                passwordless: true,
                 trustedKyc: false,
                 quoteId,
                 flow: 'buyCrypto',
@@ -496,7 +502,7 @@ export const paybisProvider: FiatProviderFactory = {
                 cryptoPaymentMethod: 'partner_controlled_with_redirect',
                 partnerUserId,
                 locale: locale.localeIdentifier.slice(0, 2),
-                passwordless: false,
+                passwordless: true,
                 trustedKyc: false,
                 quoteId,
                 flow: 'sellCrypto',
@@ -505,20 +511,25 @@ export const paybisProvider: FiatProviderFactory = {
               }
             }
 
-            const response = await paybisFetch({ method: 'POST', url, path: 'v2/public/request', apiKey, bodyParams })
-            const { requestId } = response
+            const privateKey = atob(privateKeyB64)
+            const promise = paybisFetch({ method: 'POST', url, path: 'v2/public/request', apiKey, bodyParams, privateKey, showUi })
+            const response = await showUi.showToastSpinner(lstrings.fiat_plugin_finalizing_quote, promise)
+            const { oneTimeToken, requestId } = asPublicRequestResponse(response)
 
             const widgetUrl = isWalletTestnet(coreWallet) ? WIDGET_URL_SANDBOX : WIDGET_URL
+
+            const ott = oneTimeToken != null ? `&oneTimeToken=${oneTimeToken}` : ''
+
             if (direction === 'buy') {
               await showUi.openExternalWebView({
-                url: `${widgetUrl}?requestId=${requestId}`
+                url: `${widgetUrl}?requestId=${requestId}${ott}`
               })
               return
             }
 
             const successReturnURL = encodeURI(RETURN_URL_SUCCESS)
             const failureReturnURL = encodeURI(RETURN_URL_FAIL)
-            const webviewUrl = `${widgetUrl}?requestId=${requestId}&successReturnURL=${successReturnURL}&failureReturnURL=${failureReturnURL}`
+            const webviewUrl = `${widgetUrl}?requestId=${requestId}&successReturnURL=${successReturnURL}&failureReturnURL=${failureReturnURL}${ott}`
             console.log(`webviewUrl: ${webviewUrl}`)
             let inPayment = false
 
@@ -671,13 +682,23 @@ const paybisFetch = async (params: {
   url: string
   path: string
   apiKey: string
+  showUi?: FiatPluginUi
   bodyParams?: object
   queryParams?: JsonObject
+  privateKey?: string
 }): Promise<JsonObject> => {
-  const { method, url, path, apiKey, bodyParams, queryParams = {} } = params
+  const { method, url, path, apiKey, bodyParams, queryParams = {}, privateKey, showUi } = params
   const urlObj = new URL(url + '/' + path, true)
   const body = bodyParams != null ? JSON.stringify(bodyParams) : undefined
 
+  let signature: string | undefined
+  if (privateKey != null) {
+    if (body == null) throw new Error('Paybis: Cannot sign without body')
+    // Because we will be doing a slow CPU operation in sha512HashAndSign, we need to first
+    // call waitForAnimationFrame to ensure the UI spinner is rendered.
+    if (showUi != null) await showUi.waitForAnimationFrame()
+    signature = sha512HashAndSign(body, privateKey)
+  }
   queryParams.apikey = apiKey
   urlObj.set('query', queryParams)
 
@@ -687,6 +708,13 @@ const paybisFetch = async (params: {
       'Content-Type': 'application/json'
     }
   }
+  if (signature != null) {
+    options.headers = {
+      ...options.headers,
+      'x-request-signature': signature
+    }
+  }
+
   if (body != null) {
     options.body = body
   }